final refactoring for modules
