+++ src/components/Layout.tsx
@@ -4,7 +4,7 @@ import Landing from './Landing';
 import ViewPortfolio from './ViewPortfolio/ViewPortfolio';
 import EditEmpPortfolio from './Portfolio/PortfolioEdit/EditEmpPortfolio';
 import PortfolioList from "./Portfolio/PortfolioList/PortfolioList";
-import Adminpage from "./Adminpage";
+import Adminpage from "./Admin/Adminpage";
 import Portfoliodetails from "./Portfoliodetails";
 
 function Layout() {
