+++ src/components/Portfolio/PortfolioEdit/EditEmpPortfolio.tsx
@@ -24,11 +24,11 @@ import Matrix from "../../../interfaces/Matrix";
 import CertificationContainer from "../../Certifications/CertificationContainer";
 import EducationContainer from "../../Education/EducationContainer";
 import HonorAwards from "../../Honors/HonorAward";
-import IndustryEquivalency from "../../IndustryEquivalency";
-import OtherWorkExperience from "../../OtherWorkExperience";
+import IndustryEquivalency from "../../IndustryEquivalency/IndustryEquivalency";
+import OtherWorkExperience from "../../OtherWorkExp/OtherWorkExperience";
 import Project from "../../Project";
-import AboutMe from "../../RevatureAboutMe";
-import RevatureWorkExp from "../../RevatureWorkExperience";
+import AboutMe from "../../AboutMe/RevatureAboutMe";
+import RevatureWorkExp from "../../RevatureWorkExp/RevatureWorkExperience";
 import ScrollButton from "../../ScrollButton";
 
 import { useEffect, useState } from "react";
