+++ src/test/Component.test.tsx
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 import { render, screen, waitFor } from "@testing-library/react";
 import * as rcookies from 'react-cookie';
 import { BrowserRouter, Route } from "react-router-dom";
-import Adminpage from "../components/Adminpage";
+import Adminpage from "../components/Admin/Adminpage";
 import validate from "../components/Account/FormValidation";
 import HonorAwards from "../components/Honors/HonorAward"; 
 import PortfolioList from "../components/Portfolio/PortfolioList/PortfolioList";
